+++ src/app/studies/components/StudyDescriptionCard.tsx
@@ -4,10 +4,7 @@ import Card from "@/shared/components/atoms/Card";
 import Tag from "@/shared/components/atoms/Tag";
 import Typography from "@/shared/components/atoms/Typography";
 import Meta from "@/shared/components/molecules/Meta";
-import { MDXRemote } from "next-mdx-remote-client/rsc";
-import rehypeSanitize from "rehype-sanitize";
-import rehypeRaw from "rehype-raw";
-import rehypeHighlight from "rehype-highlight";
+
 import "highlight.js/styles/github-dark.css";
 
 import ShareButton from "@/app/studies/components/applyStudy/ShareButton";
@@ -20,6 +17,7 @@ import {
   GetStudyRequirementsResponse,
   GetStudyRulesResponse,
 } from "@/features/studies/types/studies.api";
+import CustomMdxRemote from "@/shared/components/system/CustomMdxRemote";
 
 interface StudyDescriptionCardProps {
   studyDetailData: GetStudyDetailResponse;
@@ -95,31 +93,9 @@ const StudyDescriptionCard = ({
         <ContentWrapper>
           <Typography.SubTitle1>스터디 소개</Typography.SubTitle1>
           {hasItem(studyDetailData.content) && (
-            <MDXRemote
-              source={escapeCurlyBracesOutsideCodeBlocksAndInlineCode(
-                studyDetailData.content
-              )}
-              components={{
-                p: (props) => <span {...props} />,
-              }}
-              options={{
-                mdxOptions: {
-                  rehypePlugins: [
-                    [
-                      rehypeHighlight,
-                      rehypeRaw,
-                      {
-                        passThrough: ["mdxJsxFlowElement", "mdxJsxTextElement"],
-                      },
-                      rehypeSanitize,
-                    ],
-                  ],
-                },
-              }}
-            />
+            <CustomMdxRemote content={studyDetailData.content} />
           )}
         </ContentWrapper>
-
         <ContentWrapper>
           <Typography.SubTitle1>참여 요건</Typography.SubTitle1>
           {hasItem(requirementsData) ? (
@@ -180,50 +156,6 @@ const ListContent = (
 function hasItem<T>(item: T[] | T) {
   if (Array.isArray(item)) return item.length > 0;
   else return Boolean(item);
-}
-
-function escapeCurlyBracesOutsideCodeBlocksAndInlineCode(input: string) {
-  const codeBlockRegex = /```[\s\S]*?```/g; // 코드블럭 잡기
-  const inlineCodeRegex = /`[^`]*`/g; // 인라인 백틱 코드 잡기
-  const pTagRegex = /(<p>)([\s\S]*?)(<\/p>)/g; // <p> 태그 잡기
-
-  // 1. 코드블럭 먼저 임시 토큰으로 치환
-  const codeBlocks: unknown[] = [];
-  const codeBlockPlaceholder = "___CODE_BLOCK_PLACEHOLDER___";
-  let temp = input.replace(codeBlockRegex, (m) => {
-    codeBlocks.push(m);
-    return codeBlockPlaceholder;
-  });
-
-  // 2. 인라인 백틱 코드 임시 토큰으로 치환
-  const inlineCodes: string[] = [];
-  const inlineCodePlaceholder = "___INLINE_CODE_PLACEHOLDER___";
-  temp = temp.replace(inlineCodeRegex, (m) => {
-    inlineCodes.push(m);
-    return inlineCodePlaceholder;
-  });
-
-  // 3. <p> 태그 내 중괄호만 변환
-  const escaped = temp.replace(pTagRegex, (full, open, content, close) => {
-    // content 내 중괄호만 변환
-    const escapedContent = content
-      .replace(/\{/g, "&#123;")
-      .replace(/\}/g, "&#125;");
-    return open + escapedContent + close;
-  });
-
-  // 4. 인라인 코드 원복
-  let output = escaped;
-  for (const inlineCode of inlineCodes) {
-    output = output.replace(inlineCodePlaceholder, inlineCode);
-  }
-
-  // 5. 코드블럭 원복
-  for (const codeBlock of codeBlocks) {
-    output = output.replace(codeBlockPlaceholder, codeBlock as string);
-  }
-
-  return output;
 }
 
 export default StudyDescriptionCard;
